backend/requirements.txt
flask==3.1.0
flask-cors==5.0.1
pytesseract==0.3.13
Pillow==11.1.0
openpyxl==3.1.5
python-dotenv==1.1.0
google-generativeai
